More sharing of uncounted arrays
Summary:
Record whether there's a co-allocated ApcTypedValue on the ArrayData
so that we don't have to pass a flag to *::ReleaseUncounted. This
allows us to just re-use uncounted arrays in apc, rather that making
copies.

Also noticed that for collections backed by static arrays, the static
arrays were copied to uncounted arrays due to a short cut the
collections code took.

Finally noticed that although when called from the top level, there's
no possibility of any sharing in keysets (their keys must be unqiue,
and they can't contain any sub-arrays), keysets contained in other
arrays can be duplicates, or contain keys that are duplicates of
strings from elsewhere in the array; so add a PointerMap to
SetArray::MakeUncounted.

Reviewed By: ricklavoie

Differential Revision: D6951321

fbshipit-source-id: 57da1cafe6e56bfe5d057ba8bc0f86893961f4a4