Make member init order in Ast dependent on compatibility mode
Summary:
The order of member initialisations in member variables defined in the
arguments of `__construct` was reversed in D5190562, to make HackC generate
intialisation code in the same order as HHVM. For compatibility with
`Parser_hack` (and comparisons with it in the effort to cut over `hh_server`),
they should, however, not be.

This diff makes the reversal dependent on `hhvm_compat_mode`.

Reviewed By: gregory-nisbet

Differential Revision: D6922923

fbshipit-source-id: a8fba4d57da39853652c5eed561742f6e88304d8