HHVM Debugger: initialize server variables for dummy request thread
Summary:
When invoked from the command line, certain server variables (like request start time) are set in $_SERVER, but this is not happening for the dummy request thread in the new debugger. The old XDebug debugger got this for free, because the dummy was started as a separate CLI invocation of HHVM, which went through the CLI initialization path.

This is causing things like request_time() to fail when used in the console REPL.

Reviewed By: rightparen

Differential Revision: D6962253

fbshipit-source-id: 5524a1109da609b36b5e0a8feee73a949ec7fbd0