Add FCallM/RetM bytecodes
Summary:
Introduce a new set of call (FCallM, FCallDM, FCallUnpackM) and return (RetM) bytecodes to support returning multiple values via the VM stack. Currently Uninit values must be pushed to the stack to later be overridden by RetM when multiple values are returned. This is primarily useful in the JIT, where we need space allocated above the ActRec in advance for the return values (as the ActRec is expected to be partially live during the return and cannot be clobbered).

The FCallM bytecodes use a new push/pop signature (UFMANY and CMANY) to indicate that they pop the uninit values as well as the F-flavors and ActRec, and push cells. The RetM bytecode pops multiple values to return. The actual return value is the last value popped but the first value pushed. This is mostly because pushing the return value first makes the most sense at the bytecode level but for the JIT/VM it needs to be the value that overrides the base of the ActRec to match the handling of other call/return instructions in the unique stubs.

A runtime option, Eval.UseMSRVForInOut was added to enable/disable this new call ABI for inout functions at the parser and hhbbc level. It has been defaulted to true. A followup diff will ensure that both modes continue to receive test coverage.

Lastly, hhbbc tracks the return value of these functs as a vec of values. In addition to being a relatively simple way to handle multiple return values, it mirrors future plans which may allow hhbbc to mark functions as optionally returning tuples via the stack in cases where they are always destructured at the callsite.

Reviewed By: ricklavoie

Differential Revision: D6747802

fbshipit-source-id: ac4e68d3105056c020671d8d27ddc1ebb37b93ad