Make hphp/hack/test/review.sh work in Emacs shells
Summary: The script `hphp/hack/test/review.sh` behaves funny if run in Emacs shells (started by `M-x shell`).  The shell process doesn't receive any input until the user presses `RET` but then the `-n 1` seems to have the effect that `y` followed by `RET` is interpreted by bash as two characters and `RET` is received by the iteration of the loop.  So, the overall effect is that the script basically skips every other test.  The solution is not to use `-n 1` in dumb terminals.

Reviewed By: arxanas

Differential Revision: D6963801

fbshipit-source-id: 396ccb171b41853ae37f4203d18cc58866c94383