Lift Hack_format helpers to the top level
Summary:
Previously, all helpers in Hack_format were bound to local variables inside the top-level `transform` function. This made outline view useless, since it only displayed `transform`.

This change lifts all helpers to the top level so that they can be displayed in outline view. To do so, we add an `env` argument to thread the formatting environment through each helper (previously, the helpers closed over the env passed to `transform`).

Additionally, this change adds a .mli file so that only `transform` is exposed. This revealed that the `Token` module alias was unused, so we remove it.

Reviewed By: arxanas

Differential Revision: D6970805

fbshipit-source-id: dbca0ad1c0b3035780af30f7789aac25a98072ad