Enabling hash functions that require SSE4.2 support
Hash functions for HHVM are currently guarded by __SSE4_2__ flag which are not
enabled by default. The fallback versions for "hash int" functions are "C"
implementations leaving some room for performance. This change enables the SSE4.2
support to be configurable allowing all the hash functions to take the corresponding
assembly routines when ON.

This gives ~0.4% performance improvement on oss-performance/WordPress and Drupal7