Add generic build options system to typechecker
Summary:
Some currently-hardcoded paths are system-specific.

I can't just specify them in /etc/hh.conf as that path is itself system-specific - e.g. requires system integrity protection overrides on MacOS.

Once we use OPAM in open source builds, it could be replaced with ppx.

If we end up generating the FB side too, we could use this to replace the generated C code for things like build ID and versions.

Original goal of this change was for hh_server to be able to use hackfmt out of the box on open source builds.

Closes https://github.com/facebook/hhvm/pull/8120

Reviewed By: paulbiss

Differential Revision: D6929146

Pulled By: fredemmott

fbshipit-source-id: 78a54cde1f5c99566005624d70e55b95f928ff99