Introduce TDynamic
Summary:
This creates a first class dynamic type in Hack. The dynamic type is kind of like a combination of mixed and the Tany type. Like mixed it is a supertype of every other type. Like Tany any operation is valid and no errors will be reported.

What distinguishes dynamic from Tany is that dynamic is not a subtype of any other type. In other words it will only unify with itself.

The initial version of this diff is just the basics: threading through the information about the type into our NAST, and handling it on most circumstances.

In this diff, every operation with a dynamic type results in a dynamic type. The goal of this stack of diffs is to implement dynamic's interactions with other types according to specification. This diff just provides the baseline behavior.

Reviewed By: dlreeves

Differential Revision: D6764235

fbshipit-source-id: 5c1c0b314d8da2e2ebe343e9ae262380365683c5