Replace mixed with ?nonnull during the transformation of a declaration phase type into a localized type
Summary:
A third attempt at fixing the problem described in D6923539.  This time we replace `mixed` with `?nonnull` during the transformation of a declaration phase type into a localized type.  The intuition here is that `Tmixed` should exist only in the `decl` phase and should be transformed away when a type is localized.  In fact, this inspired the approach I followed to systematically find all the places that required changes:

- Changed the definition of the `Tmixed` constructor to `Tmixed : decl ty_`.
- Commented out all occurrences of `Tmixed` that according to the OCaml type checker were invalid.
- Addressed the remaining cases (added translation to `Toption (Reason.Rmixed, Tnonnull)` behind a feature flag).
- Ran the tests and confirmed that the changes to the expected outputs were reasonable and expected.
- Uncommented previously commented occurrences of `Tmixed`.  I considered failing (behind a feature flag) in such cases instead but that resulted in a lot of clutter which I decided against.

I have also chosen to introduce a new reason `Rmixed` to slightly improve error messages.

Reviewed By: andrewjkennedy

Differential Revision: D6950271

fbshipit-source-id: 6310bfc43d373943620abd9bb221b3a1e4644193