Refactor ManagedArena
Summary: The existing implementation is too tightly coupled with 1G huge pages. Here we try to change it into a general-purpose C++ wrapper of the jemalloc arena with customized allocation hooks. The detailed hook implementation is in the `ExtentAllocator` policy class. We now only have a bump allocation policy implemented.  Different page sizes are supported in different mapper classes, and we use a chain of mappers to add flexibility.  E.g., it is possible to merge the low_arena, and low_huge1g_arena, so that 1G huge pages are used when configured, and normal (or 2M) pages are used as a fallback.

Reviewed By: interwq

Differential Revision: D6894959

fbshipit-source-id: 92463ebe0f79e9832fabced1c7ff44698621aede