Reject untyped lambda in non-function-typed context in strict mode
Summary:
Currently in strict mode we miss errors in bodies of lambdas with no type hints on parameters
that are passed to functions with expected type of "mixed" or other non-function type.

This is inconsistent with our requirement that top-level functions should have type hints in
strict mode. So let's reject such lambdas. Currently this is put under an experimental flag `experimental_disallow_untyped_lambda_as_non_function_type`.

Reviewed By: jamesjwu

Differential Revision: D6981297

fbshipit-source-id: e3ff3c5ce5d51bc820498f88167facabc1f114a3