Fix escaping string literals in various cases
Summary:
This fixes two bugs:

Interpolated heredocs need to be escaped using heredoc escaping rules, not doublequotedstring rules, but interpolated heredocs are parsed as a SyntaxList, so there's no way to tell the location without looking at the token kind in the lowerer.

Second, \X42 = \x42 in PHP, so both need to escape to the letter A. This adds the "X" case.

Reviewed By: vladima

Differential Revision: D6951735

fbshipit-source-id: fbae7c3006fdc0b832f444625d972a46ed9880f6