Add runtime support for treating d/varray as dict/vec
Summary:
Add runtime support for turning darrays and varrays into dicts and vecs. This is
controlled by a new option called HackArrDVArrs. For the most part, this diff
transforms the places where darrays or varrays are created in the runtime to
either create PHP arrays or Hack arrays depending on the flag. It also adds a
number of defensive asserts to ensure that we don't encounter varrays or darrays
when this option is on (since they shouldn't exist in that case).

It also changes a varray_or_darray type-annotation to not have an associated
datatype. This is needed because vec_or_dict cannot not (because it represents
two different datatypes), thus for compatiblity varray_or_darray cannot either.

Note that this diff only adds runtime support. The parser, emitter, and HHBBC
haven't been changed, so actually turning on this option will not work. The
remaining diffs in the stack will add the remaining support.

Reviewed By: paulbiss

Differential Revision: D6970197

fbshipit-source-id: b6e550f6d67e2e735180b6f915e02ae6615fff23