Add support to the parser and emitter for HackArrDVArrs
Summary:
Add associated support for the HackArrDVArrs option to the parser and
emitter. For the most part, varray and darray are transformed to vec and dict at
the parser level. Most of the emitter support is just proactively asserting that
we never actually encounter those tokens if HackArrDVArrs is enabled, and
modifying the few places in the emitter where we create them internally.

Also add similar support to HackC. There we don't modify the parser and instead
control what bytecodes we emit directly.

With this diff, only HHBBC needs support for HackArrDVArrs.

Reviewed By: paulbiss

Differential Revision: D6970219

fbshipit-source-id: ba00c3b92849379b7f5f9758064aebde47fa108c