Fix ocamldebug issue with local module include
Summary:
Turns out that ocamldebug is having trouble with local module include, so this diff is just a workaround for that.
Because we don't see any problems running our binaries(only debugging) I would assume that the problem is connected to the fact that `ocamldebug` uses bytecode compiler while everything else compiles to native code with `ocamlopt`.

Differential Revision: D6998366

fbshipit-source-id: 315314fd412b1f217bc65319cd937afff6834532