Add support for COMPILER_HALT_OFFSET(kinda)
Summary:
COMPILER_HALT_OFFSET is defined as the offset of the end of the position of the __HALT_COMPILER() directive.

This diff implements it, but it looks like the HHVM parser actually gets this wrong when there's an interpolated string before HALT_COMPILER. With these changes, HackC gets the same result as PHP5 and 7 here.

https://3v4l.org/JSXoW

Reviewed By: vladima

Differential Revision: D6971262

fbshipit-source-id: b6457de2fa5833ab14886d0dc2ffd6b13555047c