Expose more of the parser configuration through the CLI of `hh_parse`
Summary:
jamesjwu asked to make parsing more debuggable. We've grown an extensive web of
configurable options for parsing. This diff exposes command line options to set
or unset the boolean choices of that configuration in `hh_parse`.

NOTE: It also renames the option `--hhvm_compat_mode` to `--hhvm-compat-mode`
for consistency of naming. Sorry for churn, users thereof.

Reviewed By: jamesjwu

Differential Revision: D7000494

fbshipit-source-id: 64b713d201d4e9a43c044d9402dd5a7934b2a6c1