Test case showing that typing is not idempotent
Summary:
If we want to start cancelling jobs (we do), we need to be able to restart / resume them in a way that will not loose information. The jobs unfortunately have side effects that make it not as straightforward as usage of functional language would suggest.

The approach I'm going to try here (in next diffs) is to store errors revealed when declaring X along with X in shared heap, and ensuring that they are always returned when typechecking X. This will not be strictly idempotence (since jobs can still return different sets of errors between runs), but will ensure that all the errors are always EVENTUALLY returned somwhere (right now they are gone forever).

Reviewed By: alexchow

Differential Revision: D6854116

fbshipit-source-id: 4a653a1e819be8ab78a4be770f8f2710d0e4e0df