Fix issues with cycles in store-elim
Summary:
When we find a cycle, we sometimes find a single store that can be
used, without needing to insert a phi node, but we would still mark it
as "processed" meaning that we'd constructed a store and a "magic"
Mov (with src=dst) as a placeholder to be fixed later. If the store we
found happened to be one of these special stores (which means we must
have had two intersecting cycles), then we'll try to process the magic
Mov twice, and assert the second time.

This fixes it so that resolve_cycle just creates a normal
"instruction" node, rather than a "processed" node in those cases, so
the magic Mov only gets processed in the block where it was inserted.

Reviewed By: paulbiss

Differential Revision: D6997100

fbshipit-source-id: accecf25ef655070e09672a49eca83045920d8cc