Fix isFPassStar
Summary:
FPassM was missing.

There are only two uses of isFPassStar, one goes on to assert that its
an FPassL, so isn't affected, and the other just infers a better jit
type when the instruction is an FPass - so this wasn't breaking
anything, but we might as well be complete.

Reviewed By: paulbiss

Differential Revision: D6988073

fbshipit-source-id: 741cce9eb848817b72beef5e51c464bc514271ae