Don't fatal trying to eval default value with static:: for reflection
Summary:
We can't really get the proper value in that case since static:: depends on how
the method is invoked at runtime and reflection doesn't invoke the method.
Instead of fataling whenever eval'ing the default argument fails, this diff
changes the behavior to setting the default value to null and raising a warning.

Reviewed By: paulbiss

Differential Revision: D7016286

fbshipit-source-id: d3852d1a1e5335aed197057a17545fe3dc40f91f