Speed up HackArrDVArrs implementation
Summary:
The introduction of the HackArrDVArrs option introduced a slight regression,
which has been tracked down to the changes in VArrayInit and DArrayInit. It is
apparently too expensive to check the runtime option for each operation (in
order to dispatch to the correct Hack array/PHP array implementation).

Instead, rework VArrayInit and DArrayInit to not check HackArrDVArrs
explicitly. Instead the option is only checked at construction time. The rest of
the time we always dispatch thru ArrayData's pseudo-vtable and let them select
the correct implementation. This is apparently more efficient. The main downside
of this is that we can't easily re-use the ArrayInit base class' implementations
and have to duplicate some code.

Reviewed By: paulbiss

Differential Revision: D7026822

fbshipit-source-id: d47cd04d69bbed2acde8bdc33407f75b7d41d342