Attempt to simplify instanceof types to avoid generated abstract types
Summary:
In many cases of instanceof on generic classes, it is possible to avoid using the generated type parameters (e.g. `T#1`) in the refined type. For example:
```
class C<T> extends Base<T> { }
function testit(Base<mytype> $x) {
  if ($x instanceof C) {
    // here $x can be given type C<mytype>
  }
}
```
The implementation looks at the constraints on the generated type parameters; for equality constraints, it simply uses the constrained type (e.g. `T#1 = mytype`).
A type `T#1` can be replaced by its upper bound if it is covariant and appears nowhere else in the constraints.
Likewise, it can be replaced by its lower bound if it is contravariant and appears nowhere else in the constraints.

Reviewed By: dlreeves, jamesjwu

Differential Revision: D7023720

fbshipit-source-id: a18857209229ce30de7bdc5e0285e9ad74c1bb0a