Fix set partition in lambda analyzer
Summary:
Previously there was a bug where the lambda analyzer was incorrectly partitioning the used outer variables set, so that when code gen occurred for:

```
coroutine function naming() {
  $a = coroutine ($continuation) ==> {
    $continuation = 4;
    suspend someFunction();
    $b = coroutine ($continuation) ==> {
      $continuation = 4;
      suspend someFunction();
    };
  };
}

```

it would generate two copies of `$continuation` and fail at typecheck time for lowered code.

Fixed it so that it wouldn't occur and added a test.

Reviewed By: michaeltingley

Differential Revision: D6933531

fbshipit-source-id: e153cdf1e348f5702f63087ca4600ad84aab7716