Ignore UNSAFE_EXPR in codegen scenarios
Summary: After D7041109 has landed we don't lower coroutines by default. This had impliciations on few tests that previously used to fail during coroutine transformations. Once this step is skipped tests were able to progress further - now they crawl thought lowering process and hthe cultprit is - extraction of UNSAFE_EXPR entries from the leading trivia of leading tokens. In trees with skewed shape this operation become very expensive. To unblock codegen tests we can just ignore the UNSAFE_EXPR extraction part (this is done in this diff), however ultimately we need to fix the underlying issue properly since it can become a problem for editors if one of such files is opened in the editor. As a better fix we can i.e. try to track frontier where we've already looked for trivia and bail out if node starts before the frontier

Reviewed By: oulgen, paulbiss

Differential Revision: D7052682

fbshipit-source-id: 26edbb05b71a5e2ae753aa12848b8cfdb09d3f11