Typecheck conditional reactivity
Summary:
As in title:
- Conditionally reactive methods are annotated with `__RxIfImplements`, `__RxLocalIfImplements` or `__RxShallowIfImplements`. Annotation accepts type reference T as an argument.
- Conditionally reactive method is always typechecked with restricted mutability
- When conditionally reactive method with type annotation T is called it is treated as reactive if:
  - receiver of method call is a subtype of T
  - caller of method is conditionally reactive method with type annotation T1 and T = T1

Implementation notes:
Now reactivity need to store type that will be used as reactivity criterion. Ideally reactivity should be polymorphic in 'phase so we can start with `decl ty` which then will be converted to `locl ty`:
```lang=ocaml
type 'phase reactivity = Reactive of 'phase ty option | ...
```
However doing this hits the bug in Ocaml compiler which started to crash in non-useful message `exception Ctype.Unify(_)`. As a workaround reactivity is not polymorphic and always stores `decl ty` which can later be localized.

Reviewed By: jamesjwu

Differential Revision: D7035793

fbshipit-source-id: 911d65e8383f14463635608501e5dbcb707ee48f