Handle classname typehints without type parameters
Summary:
The typechecker should prevent this, but if a user manages to use an `is` expression with a classname without a type param, we simply check if the left operand is the name of a class or interface.

We also strip the global slash from the hint name when checking if it's a classname.

Reviewed By: oulgen

Differential Revision: D7048757

fbshipit-source-id: 691e48c3f3c52d4d7d77125a28781fe3fb0b6d4f