Optimize APCHandle::toLocal
Summary:
There was a fast path for uncounted php arrays, when called from ConcurrentSharedStore, but not for any of the other typed-value like APCTypedValues.

This makes a fast path for all APCTypedValues with a valid data type. It might pessimize the php array case slightly; but should make up for it by optimizing lots more types. It also gets rid of the assumption that APCTypedValues for uncounted arrays are always co-allocated (which we need to drop if we want to share uncounted arrays).

Reviewed By: swtaarrs

Differential Revision: D7057275

fbshipit-source-id: 0bac724d64a90a7e05cd43d1ea667c213dd13fa9