Make uncounted arrays aware of co-allocated APCTypedValues
Summary: uncounted arrays can be co-allocated with an APCTypedValue. This encodes that into the uncounted array, so that the owner doesn't have to track it separately. This is important if we're going to share uncounted arrays, since there will be multiple owners, and the co-allocated one may not be the last to release it.

Reviewed By: alexeyt

Differential Revision: D7038121

fbshipit-source-id: 76d4fe7ab8dd2b3687465143e377b74fb280b670