Expose smart constructors state after the parsing
Summary:
This change will give users ability to get smart constructors state after the parsing. It's generally useful and we need it for example to examine stack in the `VerifySmartConstructors`.

There is ongoing discussion whether we want to expose a whole `sc_state` to users or just add another "compose_result" function to smart-constructors signature. But for now, it's ok to make it this way.

Differential Revision: D7026094

fbshipit-source-id: 8c96c052b29fd473f186e67a218b852099451fda