XED optimizations: caching lengths, remove xedBuff
When an offset relative to the next instruction is required, instead
of calling xedEmitPrep* twice, once to get the length of the
instruction being generated and once again to actually emit it,
the result of the first xedEmitPrep* is cached in an unordered_map.
All X64Assembler::xedEmit* methods take a pointer argument
which indicates the destination buffer where instructions will be
emitted. In most cases, codeBlock.frontier() will be passed as that
argument, and after, the call, the codeBlock's frontier is updated
with a call to the newly added DataBlock::moveFrontier.
X64Assembler::m_xedInstrBuff is removed.
Add the X64Assembler::pad() method which efficiently pads the current
X64Assembler::codeBlock with int3s and ud2s.