Fix hugifyText when dev is more than 5 chars
Summary:
I was trying to test a change to hugifyText, and it never got called
on my devserver. It turned out to be because the dev field in
proc/pid/maps was 103:01 which was too long for the sscanf
pattern. This meant that the leftover "1" was picked up as the inode,
and the inode was picked up as the filename.

Also add a warning when we don't call hugifyText from a binary that
was setup to do so.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D7085437

fbshipit-source-id: 6af07761c8b198a2d9b85a8b28c410fc87e24151