Explicitly enable/disable background threads in hugifyText
Summary:
The previous attempt uses malloc_conf to disable background threads,
and then enables them in hphp_process_init.

The problem with that is that MALLOC_CONF can still be used to enable
the background threads from the start (which leaves us with the same
problems as before), and it can no longer be used to disable
background threads, since we enable them unconditionally.

Since we only wanted to disable background threads for hugifyText, put
the code there, and save-disable-restore, rather than unconditionally disable/enable.

Reviewed By: paulbiss

Differential Revision: D7085748

fbshipit-source-id: 3b6441a1c59add6423400bb3384cfcff974a7c1d