Embed hackc binary in hhvm and ENABLE it
Summary:
This diff adds a section to the hhvm binary containing a gziped copy of hackc (it's about 8mb right now). It adds options to extract the binary to a location (default: /var/run) or fallback location (default: /tmp), and run it from there. If the hackc binary exists it can be optionally verified to ensure it is the correct version, however if the naming scheme is version specific this is unnecessary. The default name is `hackc_<repo_schema>`, and the hhvm repo schema now includes changes to the hack directory.

The zend tests are currently disabled (44 of them are still failing, and are being actively fixed), and three parser specific tests were turned off. One notable test was disabled, double-hexstring.php, it's actually broken in hhvm (the JIT isn't handling it properly), but hphpc was constant folding the expressions, masking the issue. I will put up a separate diff to address this one.

Reviewed By: swtaarrs

Differential Revision: D7083798

fbshipit-source-id: 7660e5b0043489086327543b2f1fc224cdc01ab4