HHVM Debugger: cache variables responses for server constants and globals
Summary:
Server constants, user contants and server globals are going to be the same for each request, but the client issues a VariablesRequest for them for each stack frame for each request thread, which is expensive and wasteful to compute each time.

The cached values are cleared when any request steps (since the user code could write to them) or if an evaluation is performed (since the expression could write to them).

Reviewed By: velocityboy, rightparen

Differential Revision: D6990224

fbshipit-source-id: 372df7b20a0ac3ee9cf3f8e5382b5c2e46cf2550