Handle saved_envs on lambdas in Tast_visitor
Summary: Previously, we only restored the typing environment from a function's saved_env for toplevel function definitions. This meant that we ignored the saved_env for lambda expressions. In order to make use of it, this change adds an on_fun_ override to all Tast_visitor classes, which properly sets up the typing environment for lambda expressions.

Reviewed By: pittsw

Differential Revision: D6973297

fbshipit-source-id: 82e58758fe437ef8dd025f183b21be3b01699e5d