Thread smart constructors through full fidelity parser
Summary:
This is a main diff in this stack that replaces explicit syntax-tree creation with abstract "smart constructors".
To do this I had to change underlying structure that we pass throughout the parser. This requires a lot of changes just to pass compile-time errors. Most of the changes are just `make_* -> Make.*` function conversion and changing the return type.

Thanks to `SyntaxSmartConstructors` that we are using underneath this diff **should not** change parser behavior.

Differential Revision: D6833879

fbshipit-source-id: 6acb80a57868a2e881a4d7b77d57e67043746de8