Always include EndOfFile node in the parse tree
Summary:
Full fidelity parser(FFP) included `EndOfFile` node to a syntax tree only if it had some leading trivia(comments, whitespaces at the end of the file).

FFP will always include `EndOfFile` node to the syntax tree.

We are trying to move to using smart-constructors(user provided set of functions that will be called instead of actual tree node creation) and `EndOfFile` is a great indicator for people to use as the last instruction parsed in the file.

As we verify performance impact of smart constructors, we'll try to establish more "contracts" with users on the functions we call and the order we use. Ideally, we want to have smart-constructors to be called in post-order traversal of the tree.

* Change to declaration parser to always include `EndOfFile`
* Update `hackfmt` to handle `EndOfFile` properly
* Update `full_fidelity_ast` to correctly handle `__halt_compiler`

Differential Revision: D7009377

fbshipit-source-id: 3a6a20fdcbe644400746815563447116792165cd