Split up worker.ml
Summary: This is just a minor refactor which splits up `worker.ml`. It's bugged me for awhile that `worker.ml` contained some code which ran on the worker process and some code which ran on the server process. This diff puts the server process code into `workerController.ml` and the worker process code into `worker.ml`.

Reviewed By: avikchaudhuri

Differential Revision: D6846315

fbshipit-source-id: 20a7af75fac8962683603b297a981395f57c4ac3