Run finalize_init even when typecheck_after_init is disabled
Summary:
Broken in {D7094341}. With typecheck_after_init set to false (which is the default for now), we would never call it. It's bad because we log our most important event ("init_really_end") there, as well as display "Server is READY" message, that some scripts (like memory comparison one) use to determine when server is ready (and they are getting stuck now).

(Note: this ignores all push blocking failures!)

Reviewed By: ljw1004

Differential Revision: D7114637

fbshipit-source-id: f3734ed110a5f6a83e3f298d2b42ef3796c9fd68