Flag to ban conditional expressions without a consequence node
Summary:
Provide a parser option that, when enabled, produces an error on Hack files when encountering `?<whitespace>:` in an expression context (where `<whitespace>` denotes arbitrary trivia).

PHP (since 5.3) does not treat `?:` as a separate operator but rather as a special case of the ternary operator, which allows for its middle operand to be empty. To resolve ambiguities in parsing of XHP class names (which are prefixed with `:` and may also occur in the middle part of a ternary expression), a previous change to the full-fidelity parser made `?:` (without intervening trivia) be lexed into a single token, such that the shorthand ternary is now sometimes parsed as a binary expression, other times as a conditional expression.

Between `e1 ?: e2` and `e1 ? : e2`, we now have two semantically equivalent expressions whose parse trees differ only in trivia. Preferring one over the other for formatting means failing to preserve the input tree. The parser option removes this degeneracy.

Reviewed By: jamesjwu

Differential Revision: D7091392

fbshipit-source-id: 4932aaf38a8e12ec3f79328beab4bb31e56156da